+++ internal/route/README.md
@@ -93,11 +93,14 @@ port: 443
 inbound_mtls_profile: corp-clients
 ```
 
-`TLSTermination` is valid only for TCP routes on the shared HTTPS listener. When
-enabled, Godoxy matches the TLS ClientHello SNI to the TCP route alias, completes
-TLS with the configured autocert provider, and proxies decrypted plaintext to the
-TCP target. Without it, matching HTTPS-listener TCP routes use SNI passthrough
-and the upstream target keeps serving its own certificate.
+`TLSTermination` (`tls_termination`) is valid only for TCP routes on the shared
+HTTPS listener. Route validation checks the listener address, and SNI route
+registration rejects TLS termination unless an autocert provider is configured.
+When enabled, Godoxy matches the TLS ClientHello SNI to the TCP route alias,
+terminates TLS with the configured autocert provider, and proxies decrypted
+plaintext to the TCP target. Leave `tls_termination` disabled when autocert is
+absent; matching HTTPS-listener TCP routes then remain SNI passthrough and the
+upstream target keeps serving its own certificate.
 
 ```go
 type Scheme string
