+++ internal/route/route.go
@@ -340,7 +340,7 @@ func (r *Route) validate() error {
 	if r.TLSTermination && r.Scheme != route.SchemeTCP {
 		errs.Adds("tls_termination is only supported for tcp routes")
 	}
-	if r.TLSTermination && r.Scheme == route.SchemeTCP && r.LisURL != nil && !isSharedHTTPSListenAddr(r.LisURL.Host) {
+	if r.TLSTermination && r.Scheme == route.SchemeTCP && r.LisURL != nil && !netutils.IsSharedHTTPSListenAddr(r.LisURL.Host) {
 		errs.Adds("tls_termination is only supported on the shared HTTPS listener")
 	}
 
@@ -371,25 +371,6 @@ func (r *Route) validate() error {
 		r.ExcludedReason = r.findExcludedReason()
 	}
 	return nil
-}
-
-func isSharedHTTPSListenAddr(addr string) bool {
-	host, port, err := net.SplitHostPort(addr)
-	if err != nil {
-		return addr == common.ProxyHTTPSAddr
-	}
-	httpsHost, httpsPort, err := net.SplitHostPort(common.ProxyHTTPSAddr)
-	if err != nil {
-		return addr == common.ProxyHTTPSAddr
-	}
-	if port != httpsPort {
-		return false
-	}
-	return host == httpsHost || isWildcardListenHost(host) && isWildcardListenHost(httpsHost)
-}
-
-func isWildcardListenHost(host string) bool {
-	return host == "" || host == "0.0.0.0" || host == "::"
 }
 
 func (r *Route) validateRules() error {
